move mock files under mock package
